Plan out Endpoints
We have two concerns, accounts and auth-tokens. It seems as though it
would be useful to return both things most of the time.

The only time is would make sense not to return token information is if
the request only specifies the account-id. At some point in the future,
I suspect that account-ids will become viewable to other users... who
could then steal the tokens if they knew the endpoint.

You should only be able to see the token if you're brand-new, logging
in, or already know it anyway.

I also don't think we require lookup by account-id as an endpoint yet.
It'll just be necessary when you have a token.